Demo early attempt; probably doesn't work
This is what I banged out in a night when I needed something like this.
It's either slow enough that it seems to not be working or it uses
`walkdir` incorrectly (or the library is bugged?) or it like got stuck
in a link loop on my drive or something. not sure

Anyway I don't think this will work but it might be helpful to take a
look at. Build it with `cd dedup` and `cargo build/check/run`.