Create new dirwalker
Something can potentially go wrong here: if the user supplies only
relative paths and their environment is fucked up enough that we can't
get the current working directory, we will panic. This error isn't
returned/percolated up because this is the only instance it can happen
and can never be caught (though I'm not married to this).

The code is a little funky. It relies on `collect` to transform a
Vec<Result> into a Result<Vec>