Traverse a directory
Symlinks make things weird. We should list files but not directories. We
should list symlinks, but probably only if they point to a file and not
to a directory. Rust handles both of these cases but it's not
immediately obvious. Confusingly enough DirEntry::metadata and
Path::metadata are different (the latter resolves symlinks, so it will
never return a filetype for which .is_symlink() is true).

At least the rust stdlib symlink resolver handles symlink loops.
everything's coming up milhouse