clean up. Going to add duplicated file field maybe
A `Duplicates` struct only holds links, right? Only files which have
identical IDs? Maybe it should be called Links or something? idk

So we have 2 options for storing repeated files that aren't links: to
percolate 2 (or more?) paths up from a HashProxy to the FileCatalog, or
to just store them inside the HashProxy. The latter is harder to fuck up
and easier to read but the former is faster to get information out of