Bug fixed. HashProxy can't just map Hash to Duplicates
because then collections of links get blurred/overwritten.
If A is identical to B but are not links, then A, B, all links of A, and
all links of B would end up in the same bucket (or only one set of
links, as was happening). The solution to this is a little iffy.

Uh also I did it a little wrong, this returns a vector of all the
duplicates but not organized by which are duplicates of what (just a
giant vector of every file that is repeated). Instead it should be
`get_repeats() -> Vec<Vec<(ID,Duplicates)>>` or something