remove hasher object from catalog objects
it is easier but worse to include a hasher object. They're empty structs
so I'm not sure if they take up a pointer size or not. But we should
just be able to make Catalogers generic over a FileHasher that don't
need to be initialized to be used. Then the different implementations
are added in the monomorphization step rather than being tacked on,
making it as space efficient as not using generics. Anyway, pretty sure
that's possible (and that I can do it)