switch firstkbytesproxy to use a hash type instead of bytes
before it was just using a 32-byte array of bytes. now it uses a 16-
or 32-byte hash (defined by the presence/absense of --paranoid) of the
first k bytes (defined in helpers.rs). For now K is 4096 (1 page). This
shouldn't be much slower (not really extra io, a little more
computation), etc. But we'll see

uhhh before I forget: the page is read onto the stack. I don't think
this'll cause a stack overflow? idk we'll see I guess

fixes #14